AlternatingCrossInterpolation

Documentation for AlternatingCrossInterpolation.

AlternatingCrossInterpolation.TruncationParameters

AlternatingCrossInterpolation.contract

AlternatingCrossInterpolation.elementwise

AlternatingCrossInterpolation.TruncationParameters — Type

TruncationParameters(; maxbonddimension::Int=typemax(Int), tolerance::Float64=1e-14)

A struct to hold parameters for truncation of matrices in the maxvol algorithm.

maxbonddimension: The maximum allowed bond dimension (rank) after truncation. Default is typemax(Int), which means no limit.

tolerance: The absolute tolerance for truncation. Default is 1e-14.

AlternatingCrossInterpolation.contract — Method

contract(A, Aindices, B, Bindices, indexpermutation=[])

Contract tensors A and B. The legs to be contracted are given by Aindices and Bindices.

Input:

A, B (numeric arrays): the tensors to be contracted.

Aindices, Bindices (ordered collection of integers): Indices for the legs of A and B to be contracted. The Aindices[n]-th leg of A will be contracted with the Bindices[n]-th leg of B. Aindices and Bindices should have the same size. If they are both empty, the result will be the direct product of A and B.

indexpermutation (ordered collection of integers): Permutation of of the output tensor's indices to be performed after contraction.

Output:

Contraction of A and B. All non-contracted legs are in the following order: first, all legs previously attached to A, in order, then all legs previously attached to B, in order.

AlternatingCrossInterpolation.elementwise — Method

elementwise(
    op::Function,
    inputs::Vector{<:TensorTrain{ValueType,N}};
    max_iters::Integer=20, min_iters::Integer=2,
    truncationparameters::TruncationParameters=TruncationParameters(typemax(Int), 1e-12, true),
    initial_guess::TensorTrain=randomtt(ValueType, TCI.sitedims(inputs[1]), min.([TCI.linkdims(X) for X in inputs]...))
) where {ValueType,N}

Compute the elementwise application of op to the input tensor trains in inputs using an alternating optimization procedure. The function returns a tuple containing the resulting tensor train, a vector of bond dimensions at each iteration, and a vector of pivot errors at each iteration. For example, to compute a Hadamard product of two tensor trains A and B, you could call:

result, ranks, errors = elementwise(*, [A, B])

Arguments

op::Function: The function to apply elementwise to the input tensor trains.

inputs::Vector{<:TensorTrain{ValueType,N}}: A vector of tensor trains to which the function will be applied. All tensor trains must have the same number of sites and the same local dimensions.

max_iters::Integer=20: The maximum number of iterations to perform.

min_iters::Integer=2: The minimum number of iterations to perform before checking for convergence.

truncationparameters::TruncationParameters=TruncationParameters(typemax(Int), 1e-12, true): Parameters controlling the truncation of the tensor train during the optimization process.

initial_guess::TensorTrain=randomtt(ValueType, TCI.sitedims(inputs[1]), min.([TCI.linkdims(X) for X in inputs]...)): An initial guess for the resulting tensor train. If not provided, a random tensor train with appropriate dimensions will be generated.